+++ src/main/java/app/onepass/organizer/repositories/UserOrganizationRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import app.onepass.organizer.entities.UserOrganizationEntity;
 
-public interface UserOrganizationRepository extends JpaRepository<UserOrganizationEntity, Long> {
+public interface UserOrganizationRepository extends JpaRepository<UserOrganizationEntity, Integer> {
 
 	UserOrganizationEntity findByUserIdAndOrganizationId(long userId, long organizationId);
 }
